Attempt to incorporate @ilayn's comments
"I think being a maintainer is a bit more long-term and involved (still of course anyone can become) so we might not want to paint the wrong picture about that here." - @ilayn
"You can start with the definition of a maintainer and then switch to SciPy specific details. Also approximately 15 is in fact exactly 15" @ilayn
I didn't want to lead with a separate definition; I think it is more compact and immediately relevant if the definition is provided in-line. I do prefer a colon over the em dash, though, if only to ovoid overuse of the em dash.