Swap subsections in Project Organization (#128)
* Swap subsections in Project Organization
The "Governance" and "Maintainers and contributors" subsections are focused on SciPy itself (internal). "Commmunity beyond the SciPy library" is about relationships with other projects (external), but its sandwiched between the other two internal sections. I changed the order of subsections to "Governance", "Maintainers and contributors", and "Community beyond the SciPy library"; they are now in order from closest to the core of the project to further away.

* Attempt to incorporate @ilayn's comments

"I think being a maintainer is a bit more long-term and involved (still of course anyone can become) so we might not want to paint the wrong picture about that here." - @ilayn
"You can start with the definition of a maintainer and then switch to SciPy specific details. Also approximately 15 is in fact exactly 15" @ilayn
I didn't want to lead with a separate definition; I think it is more compact and immediately relevant if the definition is provided in-line. I do prefer a colon over the em dash, though, if only to ovoid overuse of the em dash.