Reductions to paragraphs on local and global optimizers (#130)
* Reductions to paragraphs on local and global optimizers
As discussed in #113, #102
* Updates to address @apbard's comments
* Implement @apbard suggestions
"I also changed the several methods of finding local to several methods for finding local, isn't this more correct?
Do you think something like The \texttt{minimize} provides a common (or shared) interface for several methods of... solves the issue? If not, your version it is fine."
* Update per @andyfaff comments
"The new _diffev_ function is a ..."
I think this is a matter of preference, but easy enough to make the change.